requirements.txt
flask-cors
flask
numpy
speedtest-cli
yt-dlp